product add/edit/delete functions

+++ src/service/inventory-service.js
@@ -15,6 +15,51 @@ Product.belongsTo(Inventory, { foreignKey: "id" });
 Inventory.belongsTo(Store, { foreignKey: "store_id" });
 Store.belongsTo(Inventory, { foreignKey: "id" });
 
+const inventoryProductById = ({req, res, next}) => {
+    const product_id = req.params.id;
+
+    const schema = Joi.object().keys({
+        product_id: Joi.number()
+    });
+
+    const validate = schema.validate(req.body);
+
+    if(validate.error) {
+        return res.status(401).json({
+            error: validate.error.details[0].message
+        });
+    }
+
+    Inventory.findAndCountAll({
+        include: [
+            {
+              model: Product,
+              attributes: ["name", "price", "description"],
+            },
+            {
+                model: Store,
+                attributes: ["name"]
+            }
+        ],
+        attributes: ["id", "store_id", "product_id", "stock", "created"],
+        where: {
+            product_id: product_id
+        },
+        order: [["id", "DESC"]]
+    })
+    .then(inventory => {
+        res.status(200).json({
+            data: inventory
+        });
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).json({
+            error: err
+        });
+    });
+};
+
 const inventoryStoreByIdPages = ({req, res, next}) => {
     const store_id = req.params.id;
     const search = req.query.search;
@@ -43,7 +88,7 @@ const inventoryStoreByIdPages = ({req, res, next}) => {
         include: [
             {
               model: Product,
-              attributes: ["name", "description"],
+              attributes: ["name", "price", "description"],
               where: {
                   name: {
                       [Op.like]: `%${search}%`
@@ -109,7 +154,7 @@ const inventoryStoreByIdAdd = ({req, res, next}) => {
         include: [
             {
                 model: Product,
-                attributes: ["name", "description"],
+                attributes: ["name", "price", "description"],
                 where: {
                     id: inventoryData.product_id
                 }
@@ -162,6 +207,43 @@ const inventoryStoreByIdAdd = ({req, res, next}) => {
     });
 };
 
+const inventoryStoreByIdUpdate = ({req, res, next}) => {
+    const inventoryId = req.params.id;
+    const inventoryData = {
+        stock: req.body.stock
+    };
+
+    const schema = Joi.object().keys({
+        stock: Joi.number()
+            .min(1)
+            .required(),
+    });
+
+    const validate = schema.validate(req.body);
+    if (validate.error) {
+        return res.status(401).json({
+            error: validate.error.details[0].message
+        });
+    }
+
+    Inventory.update(inventoryData, {
+        where: {
+            id: inventoryId
+        }
+    })
+    .then(inventory => {
+        res.status(200).json({
+            status: "Update successful!"
+        });
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).json({
+            error: err
+        });
+    });
+};
+
 const inventoryStoreByIdDelete = ({req, res, next}) => {
     const deleteId = req.params.id;
     const schema = Joi.object().keys({
@@ -176,15 +258,34 @@ const inventoryStoreByIdDelete = ({req, res, next}) => {
         });
     }
 
-    Inventory.destroy({
+    Inventory.findOne({
         where: {
             id: deleteId
         }
     })
     .then(inventory => {
-        res.status(200).json({
-            status: "Delete successful!"
-        });
+        if(inventory) {
+            Inventory.destroy({
+                where: {
+                    id: deleteId
+                }
+            })
+            .then(inventory => {
+                res.status(200).json({
+                    status: "Delete successful!"
+                });
+            })
+            .catch(err => {
+                console.log(err);
+                res.status(500).json({
+                    error: err
+                });
+            });
+        } else {
+            res.status(401).json({
+                error: "Record not found!"
+            });
+        }
     })
     .catch(err => {
         console.log(err);
@@ -194,6 +295,8 @@ const inventoryStoreByIdDelete = ({req, res, next}) => {
     });
 };
 
+module.exports.inventoryProductById = inventoryProductById;
 module.exports.inventoryStoreByIdPages = inventoryStoreByIdPages;
 module.exports.inventoryStoreByIdAdd = inventoryStoreByIdAdd;
 module.exports.inventoryStoreByIdDelete = inventoryStoreByIdDelete;
+module.exports.inventoryStoreByIdUpdate = inventoryStoreByIdUpdate;
